reformatting

+++ src/energysys_components/energy_storage.py
@@ -20,10 +20,12 @@ class ESCParameter:
     E_in_sd2_type: ECarrier
     E_out_type: ECarrier
 
+    # Operation
     eta: float  # Const. or output dependend efficiency [-]
-    C_rate: float  # This is the charge per hour rate –
+    C_rate: float  # This is the charge per hour rate [-] –
     # one divided by the number of hours to charge the battery fully.
 
+    # Techno-economic
     spec_invest_cost: float  # [€/kWh]
     spec_volume: float  # [m^^3/kWh]
     spec_mass: float  # [kg/kWh]
@@ -60,7 +62,7 @@ class EnergyStorageComponent:
 
     def __init__(self,
                  par: ESCParameter,
-                 ts: int,
+                 ts: float,
                  state: ESCState = ESCState()):
         """
         :param par: ESCParams object
